Separate implicit and default action buttons
